requirements.txt
PyQt5
qt_material
pyserial